# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User App API",
        "version": "0.1"
    },
    "paths": {
        "/api/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\UserController::create",
                "requestBody": {
                    "description": "Request body for User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Kevin",
                                    "email": "kevin@kevin.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\UserController::readByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\UserController::read",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/user/{id}/update": {
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\UserController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Request body for User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Kevin"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/user/{id}/delete": {
            "delete": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\UserController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/user-role/create": {
            "post": {
                "tags": [
                    "User Role"
                ],
                "operationId": "App\\Http\\Controllers\\UserRoleController::create",
                "requestBody": {
                    "description": "Request body for User Role",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "user_id": "1",
                                    "position": "CEO"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/user-role/{id}": {
            "get": {
                "tags": [
                    "User Role"
                ],
                "operationId": "App\\Http\\Controllers\\UserRoleController::readByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user role",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/user-roles": {
            "get": {
                "tags": [
                    "User Role"
                ],
                "operationId": "App\\Http\\Controllers\\UserRoleController::read",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/user-role/{id}/update": {
            "put": {
                "tags": [
                    "User Role"
                ],
                "operationId": "App\\Http\\Controllers\\UserRoleController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user role",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Request body for User Role",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "position": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "position": "CEO",
                                    "status": "Active"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/user-role/{id}/delete": {
            "delete": {
                "tags": [
                    "User Role"
                ],
                "operationId": "App\\Http\\Controllers\\UserRoleController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user role",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}